==> Codes/1_geo_ml_simulation.rmd <==
# Set up

## Packages
```{r }

# /*===========================================================
#' # Preparation
# /*===========================================================
#* set up python environment
library(reticulate)

#* packages
library(sp)
library(spdep)
library(spatialreg)
library(sf)
library(raster)
library(data.table)
library(tidyverse)
library(dplyr)
library(magrittr)
library(gstat)
library(GWmodel)
library(scam)
library(mgcv)
library(magic)
library(stringr)
library(ggplot2)
library(tictoc)
library(here)
```

## Other preparations 
```{r }
#* set working directory
setwd(here())

#* source all the R functions in the Functions/R folder
fs::dir_ls(here("Codes", "Functions", "R"), full.names = TRUE) %>%
  purrr::map(~ source(.))

#* source all the python functions in the Functions/Python folder
# source_python(here("Codes", "Functions", "Python"))
source_python(here("Codes", "Functions", "Python", "run_DML_OF.py"))
source_python(here("Codes", "Functions", "Python", "run_DR_OF.py"))
source_python(here("Codes", "Functions", "Python", "run_DML_OF_c.py"))

```

# Simulation Analysis

## Prepare data
```{r }
#* Read field data
field_data <-
  readRDS(here("Data/field_data.rds")) %>%
  pull(field_sf) %>%
  .[[1]]

#* Read field parameters
field_parameters <- readRDS(here("Data/field_parameters.rds"))

#* number of simulation cases
nsim <- field_parameters$field_pars[[1]][, max(sim)]

#* Read field data
field_with_design <- readRDS(here("Data/field_with_design.rds"))

#* Read the simulated data
sim_data <-
  field_with_design$data_file_name %>%
  readRDS()

```

## Find true optimal N

```{r }
#* define prices
pN <- sim_data$pN
pCorn <- sim_data$pCorn

#* find true EONR
field_pars <-
  field_parameters$field_pars %>%
  rbindlist() %>%
  .[, opt_N := (pN / pCorn - b1) / (2 * b2)] %>%
  .[, opt_N := pmin(Nk, opt_N)] %>%
  .[, opt_N := pmax(0, opt_N)]
```

Quick visualiation of site-specific EONR

```{r }
sim_i <- 1
temp <-
  left_join(field_data, field_pars[sim == sim_i, ], by = "cell_id")

# ggplot(data = temp) +
#   geom_sf(aes(fill = Nk), color = NA) +
#   scale_fill_viridis_c()
```

## Define simulation function

```{r }
run_analysis <- function(sim_i, x_vars, models_data, pN, pCorn, nsim) {
  print(sim_i)

  # /*+++++++++++++++++++++++++++++++++++
  #' ## Prepare data
  # /*+++++++++++++++++++++++++++++++++++

  #* extract the data for the sim_i
  data <- sim_data$reg_data[[1]][sim == sim_i, ]

  #* raw train data
  train_data <-
    data$data[[1]] %>%
    .[, x := (X - min(X)) / (max(X) - min(X))] %>%
    .[, y := (Y - min(Y)) / (max(Y) - min(Y))] %>%
    .[, xy := x * y] %>%
    .[, N2 := N^2]

  #* scale data
  scaler_data <-
    train_data[, c("aunit_id", "yield", "N", "N2", x_vars), with = FALSE] %>%
    melt(id.vars = "aunit_id") %>%
    nest_by(variable) %>%
    mutate(scaler = 1 / max(data$value)) %>%
    data.table() %>%
    .[, .(variable, scaler)]

  #* scale the train data
  scaled_train_data <-
    train_data[, c("aunit_id", "yield", "N", "N2", x_vars), with = FALSE] %>%
    scale_data(., scaler_data)

  #* define parameters
  N_levels <- data$N_levels[[1]]

  #* scaled test data
  N_eval_data <-
    data.table(
      N = seq(min(N_levels), max(N_levels), length = 50)
    ) %>%
    .[, N2 := N^2]

  scaled_test_data <-
    sim_data$reg_data[[1]] %>%
    .[sim == ifelse(sim_i + 1 > nsim, 1, sim_i + 1), ] %>%
    .$data %>%
    .[[1]] %>%
    .[, c("aunit_id", "yield", "N", "N2", x_vars), with = FALSE] %>%
    .[, `:=`(
      N = NULL,
      N2 = NULL
    )] %>%
    expand_grid_df(., N_eval_data) %>%
    scale_data(., scaler_data)

  scaled_prices <-
    data.table(
      yield = pCorn,
      N = pN
    ) %>%
    scale_data(., scaler_data, back = TRUE)

  results <-
    models_data %>%
    .[on == TRUE, ] %>%
    rowwise() %>%
    mutate(results = list(
      run_indiv_analysis(
        model = model,
        train_data = scaled_train_data,
        test_data = scaled_test_data,
        x_vars = x_vars,
        prices = scaled_prices
      )
    )) %>%
    mutate(results = list(
      mutate(results, model = model)
    )) %>%
    pull(results) %>%
    rbindlist() %>%
    scale_data(., scaler_data, back = TRUE)

  return(results)
}
```


## Run simulation

 
```{r }
models_data <-
  data.table(
    model = c(
      "lm", # linear quadratic
      "gwr_t", # gwr-trevisan
      "gwr_semi", # gwr-semiparametric
      "gwr_zone_scam", # scam by zone,
      "ma_cf", # multiarm-CF
      "brf", # boosted RF
      "dmlof_semi", # semiparmetric DML-OF
      "dmlof_quad", # semiparmetric DML-OF
      "drof" # DR-OF
    ),
    on = c(
      TRUE, # lm
      FALSE, # gwr_t
      FALSE, # gwr_semi
      FALSE, # gwr_zone_scam
      FALSE, # ma_cf
      FALSE, # brf
      FALSE, # dmlof_semi
      FALSE, # dmlof_quad
      FALSE # drof
    )
  )

x_vars <-
  c(
    "b0_1", "b0_2",
    "Nk_2_1", "Nk_2_2", "Nk_1_1", "Nk_1_2",
    "plateau_2_1", "plateau_2_2", "plateau_1_1", "plateau_1_2",
    "theta_plateau_1", "theta_plateau_2",
    "theta_Nk_1", "theta_Nk_2",
    "theta_b0_1", "theta_b0_2"
  )

sim_range <- 1:20

eonr_results <-
  lapply(
    sim_range,
    function(x) run_analysis(sim_i = x, x_vars, models_data, pN, pCorn, nsim)
  ) %>%
  rbindlist()

# eonr_results[, .N, by = .(model, sim, aunit_id)] %>%
#   .[, .N] %>%
#   hist()

profit <-
  eonr_results %>%
  nest_by(model) %>%
  mutate(
    profit =
      field_pars[sim %in% sim_range, ] %>%
        data.table(field_data)[, .(cell_id, aunit_id)][., on = "cell_id"] %>%
        data.table(data)[., , on = c("sim", "aunit_id")] %>%
        .[, yield := gen_yield_QP(b0, b1, b2, Nk, opt_N_hat)] %>%
        .[, profit := pCorn * yield - pN * opt_N_hat] %>%
        .[, mean(profit, na.rm = TRUE)]
  )
```

```{r }
eonr_results <-
  lapply(
    1:50,
    function(x) run_analysis(sim_i = x, pN, pCorn)
  ) %>%
  rbindlist()

eonr_results <- readRDS("Results/eonr_results_1_50.rds")

field_tau <-
  field_pars[sim %in% 1:50, ] %>%
  data.table(field_data)[, .(cell_id, aunit_id)][., on = "cell_id"] %>%
  unique(eonr_results, by = c("sim", "aunit_id"))[., on = c("sim", "aunit_id")]
```

## Profit assessment

```{r }
field_tau %>%
  .[, .(
    cell_id, b0, b1, b2, Nk, opt_N,
    opt_N_drof,
    opt_N_drof_smth,
    # opt_N_gwr,
    # opt_N_scam,
    # opt_N_gwr_semi,
    opt_N_dmlof_c,
    opt_N_brf,
    opt_N_macf
  )] %>%
  melt(id.var = c("cell_id", "b0", "b1", "b2", "Nk")) %>%
  .[!is.na(value), ] %>%
  .[, yield := gen_yield_QP(b0, b1, b2, Nk, value)] %>%
  .[, profit := pCorn * yield - pN * value] %>%
  .[, mean(profit), by = variable]
```

## Visualization of True and Estimated EONR

```{r }
dplyr::select(
  field_tau,
  opt_N,
  opt_N_drof,
  # opt_N_drof_smth,
  opt_N_gwr,
  # opt_N_scam,
  # opt_N_gwr_semi,
  opt_N_dmlof_c,
  opt_N_brf,
  opt_N_macf,
  geometry
) %>%
  melt(id.var = "geometry") %>%
  st_as_sf() %>%
  ggplot(data = .) +
  geom_sf(aes(fill = value), color = NA) +
  facet_grid(variable ~ .) +
  scale_fill_viridis_c()
```

```{r }
ggplot(field_tau) +
  geom_point(aes(y = opt_N, x = opt_N_dmlof_c)) +
  geom_abline(slope = 1, color = "red") +
  coord_equal()

```
